+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- Option to choose transitionStyle. Set to ``undefined`` removes transition.
+  [thet]
+
 - When opening modal, set the documents body to ``overflow: hidden``, so that document's scrollbars are hidden.
   Restore, when closing the overlay.
   [thet]
-